+++ application/forms/IcingaUserGroupForm.php
@@ -2,10 +2,9 @@
 
 namespace Icinga\Module\Director\Forms;
 
-use Icinga\Module\Director\Web\Form\DirectorObjectForm;
 use ipl\Web\Url;
 
-class IcingaUserGroupForm extends DirectorObjectForm
+class IcingaUserGroupForm extends IcingaGroupForm
 {
     /**
      * @throws \Zend_Form_Exception
@@ -25,26 +24,6 @@ class IcingaUserGroupForm extends DirectorObjectForm
              ->addZoneElements()
              ->groupMainProperties()
              ->setButtons();
-    }
-
-    /**
-     * @return $this
-     * @throws \Zend_Form_Exception
-     */
-    protected function addZoneElements()
-    {
-        $this->addZoneElement(true);
-        $this->addDisplayGroup(['zone_id'], 'clustering', [
-            'decorators' => [
-                'FormElements',
-                ['HtmlTag', ['tag' => 'dl']],
-                'Fieldset',
-            ],
-            'order'  => self::GROUP_ORDER_CLUSTERING,
-            'legend' => $this->translate('Zone settings')
-        ]);
-
-        return $this;
     }
 
     protected function addAssignmentElements()
