+++ library/Director/Objects/IcingaHostGroup.php
@@ -6,21 +6,6 @@ class IcingaHostGroup extends IcingaObjectGroup
 {
     protected $table = 'icinga_hostgroup';
 
-    protected $defaultProperties = [
-        'id'            => null,
-        'uuid'          => null,
-        'object_name'   => null,
-        'object_type'   => null,
-        'disabled'      => 'n',
-        'display_name'  => null,
-        'assign_filter' => null,
-        'zone_id'       => null,
-    ];
-
-    protected $relations = [
-        'zone' => 'IcingaZone',
-    ];
-
     /** @var HostGroupMembershipResolver */
     protected $hostgroupMembershipResolver;
