+++ library/Director/Objects/IcingaObjectGroup.php
@@ -20,6 +20,11 @@ abstract class IcingaObjectGroup extends IcingaObject implements ExportInterface
         'disabled'      => 'n',
         'display_name'  => null,
         'assign_filter' => null,
+        'zone_id'       => null,
+    ];
+
+    protected $relations = [
+        'zone' => 'IcingaZone',
     ];
 
     protected $memberShipShouldBeRefreshed = false;
