+++ library/Director/Objects/IcingaUserGroup.php
@@ -6,25 +6,8 @@ class IcingaUserGroup extends IcingaObjectGroup
 {
     protected $table = 'icinga_usergroup';
 
-    protected $uuidColumn = 'uuid';
-
     /** @var UserGroupMembershipResolver */
     protected $userGroupMembershipResolver;
-
-    protected $defaultProperties = [
-        'id'            => null,
-        'uuid'          => null,
-        'object_name'   => null,
-        'object_type'   => null,
-        'disabled'      => 'n',
-        'display_name'  => null,
-        'zone_id'       => null,
-        'assign_filter' => null
-    ];
-
-    protected $relations = [
-        'zone' => 'IcingaZone',
-    ];
 
     protected function prefersGlobalZone()
     {
